+++ backend/prisma.yml
@@ -2,7 +2,9 @@ endpoint: http://localhost:4466/default/default
 datamodel: datamodel.prisma
 generate:
   - generator: typescript-client
-    output: ./generated/prisma-client/
+    output: ./generated/prisma-client
+  - generator: graphql-schema
+    output: ./generated/schema.graphql
 hooks:
   post-deploy:
     - npm run generate
